feat(server): split product endpoint
