+++ server/product/controller.go
@@ -46,51 +46,6 @@ func (ct *Controller) getById(c *echo.Context) error {
 	return c.JSON(200, result)
 }
 
-func (ct *Controller) add(c *echo.Context) error {
-	payload := AddRequest{}
-	if err := core.BindAndValidate(c, &payload); err != nil {
-		return err
-	}
-
-	user := core.GetUserSession(c)
-	result, err := ct.service.Add(c.Request().Context(), user.ID, payload)
-	if err != nil {
-		return err
-	}
-
-	return c.JSON(200, result)
-}
-
-func (ct *Controller) update(c *echo.Context) error {
-	payload := UpdateRequest{}
-	if err := core.BindAndValidate(c, &payload); err != nil {
-		return err
-	}
-
-	user := core.GetUserSession(c)
-	result, err := ct.service.Update(c.Request().Context(), user.ID, payload)
-	if err != nil {
-		return err
-	}
-
-	return c.JSON(200, result)
-}
-
-func (ct *Controller) delete(c *echo.Context) error {
-	payload := DeleteRequest{}
-	if err := core.BindAndValidate(c, &payload); err != nil {
-		return err
-	}
-
-	user := core.GetUserSession(c)
-	err := ct.service.Delete(c.Request().Context(), user.ID, payload.ID)
-	if err != nil {
-		return err
-	}
-
-	return c.JSON(200, core.CreateActionResponse(true))
-}
-
 func (ct *Controller) rent(c *echo.Context) error {
 	payload := RentRequest{}
 	if err := core.BindAndValidate(c, &payload); err != nil {
@@ -125,10 +80,7 @@ func (ct *Controller) RegisterRoutes(g *echo.Group) {
 	g.Use(core.NewGuardRoleMiddleware(core.GuardRoleUser))
 
 	g.GET("", ct.list)
-	g.POST("", ct.add)
 	g.GET("/:id", ct.getById)
-	g.PUT("/:id", ct.update)
-	g.DELETE("/:id", ct.delete)
 	g.POST("/:id/rent", ct.rent)
 	g.GET("/:id/reviews", ct.listReview)
 }
